Adding place cursor script demo file
Javascript function creates a variable, selects an input element with a particular ID in the dom then assigns that reference to a variable, finally we call the focus() method on the variable. An event listener is added to the window object which calls the placeCursor function after the load event.